+++ app.py
@@ -10,7 +10,7 @@ Functions:
     health_check(): A route to check if the service is running.
 """
 
-__version__ = "0.3.1"
+__version__ = "0.3.2"
 
 
 import logging
@@ -53,9 +53,10 @@ def step(steps: int):
         - **200 OK**: "Moved 0 steps."
     """
     for _ in range(steps):
+        elevator.update()
         logger.debug(
             f"After this step, the elevator is now at {elevator.current_floor} and "
-            "has a queue of stops for these floors: {elevator.stop_queue}."
+            f"has a status of {elevator.status} and a queue of stops for these floors: {elevator.stop_queue}."
         )
         # pylint: disable=expression-not-assigned
         [
@@ -63,13 +64,13 @@ def step(steps: int):
                 f"Currently, there are: {len(v)} persons in the location of {k}"
             )
             for k, v in elevator.persons.items()
+            if v
         ]
         # pylint: enable: expression-not-assigned
-        elevator.update()
 
     logger.info(
-        f"After {steps} steps, the elevator is now at {elevator.current_floor} and "
-        "has a queue of stops for these floors: {elevator.stop_queue}."
+        f"After {steps} step(s), the elevator is now at {elevator.current_floor} and "
+        f"has a status of {elevator.status} and a queue of stops for these floors: {elevator.stop_queue}."
     )
     return Response(f"Moved {steps} steps.", status=200)
 
@@ -102,8 +103,6 @@ def create_person():
     Body:
         JSON list of persons to add, each person must follow the format of {"origin": int , "destination": int},
         with optional keys of {"weight": float, "cargo": float}.
-
-
 
     Responses:
         - **200 OK**: